Move iocore/cache/Doc.h to iocore/cache/P_Doc.h
