+++ src/iocore/cache/Cache.cc
@@ -22,8 +22,8 @@
  */
 
 #include "iocore/cache/Cache.h"
-#include "iocore/cache/Doc.h"
 
+#include "P_CacheDoc.h"
 // Cache Inspector and State Pages
 #include "P_CacheTest.h"
