+++ src/iocore/cache/CacheWrite.cc
@@ -22,9 +22,9 @@
  */
 
 #include "P_Cache.h"
+#include "P_CacheDoc.h"
 #include "iocore/cache/AggregateWriteBuffer.h"
 #include "iocore/cache/CacheEvacuateDocVC.h"
-#include "iocore/cache/Doc.h"
 
 #define UINT_WRAP_LTE(_x, _y) (((_y) - (_x)) < INT_MAX)  // exploit overflow
 #define UINT_WRAP_GTE(_x, _y) (((_x) - (_y)) < INT_MAX)  // exploit overflow
