+++ src/iocore/cache/P_CacheDir.h
@@ -23,8 +23,9 @@
 
 #pragma once
 
+#include "P_CacheDoc.h"
+
 #include "iocore/cache/CacheDefs.h"
-#include "iocore/cache/Doc.h"
 #include "P_CacheHttp.h"
 #include "iocore/eventsystem/EventSystem.h"
 #include "iocore/eventsystem/Continuation.h"
