+++ src/iocore/cache/P_CacheTest.h
@@ -24,7 +24,7 @@
 #pragma once
 
 #include "P_Cache.h"
-#include "iocore/cache/Doc.h"
+#include "P_CacheDoc.h"
 #include "proxy/RegressionSM.h"
 
 #define MAX_HOSTS_POSSIBLE    256
