+++ src/iocore/cache/Stripe.cc
@@ -22,8 +22,8 @@
  */
 
 #include "iocore/cache/Cache.h"
-#include "iocore/cache/Doc.h"
 #include "P_CacheDisk.h"
+#include "P_CacheDoc.h"
 #include "P_CacheInternal.h"
 #include "P_CacheVol.h"
